chore: move function to shared util class
